feat: :sparkles:

+++ day-69-starting-files-blog-with-users/main.py
@@ -3,7 +3,7 @@ from flask import Flask, abort, render_template, redirect, url_for, flash,reques
 from flask_bootstrap import Bootstrap5
 from flask_ckeditor import CKEditor
 from flask_gravatar import Gravatar
-from flask_login import UserMixin, login_user, LoginManager, current_user, logout_user
+from flask_login import UserMixin, login_user, LoginManager, current_user, logout_user,login_required
 from flask_sqlalchemy import SQLAlchemy
 from functools import wraps
 from werkzeug.security import generate_password_hash, check_password_hash
@@ -41,24 +41,29 @@ db.init_app(app)
 
 
 # CONFIGURE TABLES
+class UserDb(db.Model, UserMixin):
+    __tablename__ = "users"
+    id = db.Column(db.Integer, primary_key=True)
+    email = db.Column(db.String(250), unique=True, nullable=False)
+    password = db.Column(db.String(250), nullable=False)
+    name = db.Column(db.String(250), nullable=False)
+    posts = db.relationship("BlogPost", back_populates="author")
+
 class BlogPost(db.Model):
     __tablename__ = "blog_posts"
     id = db.Column(db.Integer, primary_key=True)
     title = db.Column(db.String(250), unique=True, nullable=False)
     subtitle = db.Column(db.String(250), nullable=False)
+    author_id = db.Column(db.Integer, db.ForeignKey("users.id"))
     date = db.Column(db.String(250), nullable=False)
     body = db.Column(db.Text, nullable=False)
-    author = db.Column(db.String(250), nullable=False)
+    author = db.relationship("UserDb", back_populates="posts")
     img_url = db.Column(db.String(250), nullable=False)
 
 
 # TODO: Create a User table for all your registered users. 
 
-class UserDb(db.Model,UserMixin):
-    id = db.Column(db.Integer, primary_key=True)
-    email = db.Column(db.String(250), unique=True, nullable=False)
-    password = db.Column(db.String(250), nullable=False)
-    name = db.Column(db.String(250), nullable=False)
+
 
 login_manager = LoginManager()
 login_manager.init_app(app)
@@ -92,7 +97,7 @@ def register():
             else:
                 return "There was some error while adding the database"
 
-    return render_template("register.html",form=form)
+    return render_template("register.html",form=form,current_user=current_user)
 
 
 # TODO: Retrieve a user from the database based on their email. 
@@ -107,33 +112,37 @@ def login():
             if wanted_object:
                 if wanted_object.email==emaill and wanted_object.password==passwordd:
                     login_user(wanted_object)
+             
             
-                    return redirect("post.html")
+                    return redirect(url_for('get_all_posts'))
                 else:
-                    return "LOGIN WAS NOT A SUCCESS"
+                    flash("The password was incorrect", "error")
             else:
-                return "USER EMAIL NOT FOUND"
+                flash("The email was not found", "error")
     
-    return render_template("login.html",form=form)
+    return render_template("login.html",form=form,current_user=current_user)
 
-
-@app.route('/logout')
+@app.route('/logout',methods=["POST","GET"])
 def logout():
+    logout_user()
     return redirect(url_for('get_all_posts'))
 
 
 @app.route('/')
+@login_required
 def get_all_posts():
     result = db.session.execute(db.select(BlogPost))
     posts = result.scalars().all()
-    return render_template("index.html", all_posts=posts)
+    print(current_user.id)
+    return render_template("index.html", all_posts=posts,current_user=current_user)
+
 
 
 # TODO: Allow logged-in users to comment on posts
 @app.route("/post/<int:post_id>")
 def show_post(post_id):
     requested_post = db.get_or_404(BlogPost, post_id)
-    return render_template("post.html", post=requested_post)
+    return render_template("post.html", post=requested_post,current_user=current_user)
 
 
 # TODO: Use a decorator so only an admin user can create a new post
@@ -141,17 +150,20 @@ def show_post(post_id):
 def add_new_post():
     form = CreatePostForm()
     if form.validate_on_submit():
-        new_post = BlogPost(
-            title=form.title.data,
-            subtitle=form.subtitle.data,
-            body=form.body.data,
-            img_url=form.img_url.data,
-            author=current_user,
-            date=date.today().strftime("%B %d, %Y")
-        )
-        db.session.add(new_post)
-        db.session.commit()
-        return redirect(url_for("get_all_posts"))
+        if current_user.id==1:
+            new_post = BlogPost(
+                title=form.title.data,
+                subtitle=form.subtitle.data,
+                body=form.body.data,
+                img_url=form.img_url.data,
+                author=current_user,
+                date=date.today().strftime("%B %d, %Y")
+            )
+            db.session.add(new_post)
+            db.session.commit()
+            return redirect(url_for("get_all_posts"))
+        else:
+            return "NOT AUTHORIZED"
     return render_template("make-post.html", form=form)
 
 
@@ -174,26 +186,29 @@ def edit_post(post_id):
         post.body = edit_form.body.data
         db.session.commit()
         return redirect(url_for("show_post", post_id=post.id))
-    return render_template("make-post.html", form=edit_form, is_edit=True)
+    return render_template("make-post.html", form=edit_form, is_edit=True,current_user=current_user)
 
 
 # TODO: Use a decorator so only an admin user can delete a post
 @app.route("/delete/<int:post_id>")
 def delete_post(post_id):
-    post_to_delete = db.get_or_404(BlogPost, post_id)
-    db.session.delete(post_to_delete)
-    db.session.commit()
-    return redirect(url_for('get_all_posts'))
+    if current_user.id==1:
+        post_to_delete = db.get_or_404(BlogPost, post_id)
+        db.session.delete(post_to_delete)
+        db.session.commit()
+        return redirect(url_for('get_all_posts'))
+    else:
+        return "NOT AUTHORIZED"
 
 
 @app.route("/about")
 def about():
-    return render_template("about.html")
+    return render_template("about.html",current_user=current_user)
 
 
 @app.route("/contact")
 def contact():
-    return render_template("contact.html")
+    return render_template("contact.html",current_user=current_user)
 
 
 if __name__ == "__main__":
